api/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Geospatial
geopandas==0.14.3
shapely==2.0.2
h3==3.7.6
pyproj==3.6.1

# Embeddings (local BGE model)
sentence-transformers==2.3.1
torch==2.2.0

# LLM (optional - for synthesis)
openai>=1.40.0
httpx>=0.27.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
numpy==1.26.4

# Geocoding
geopy==2.4.1

